Updated to use django custom user model version of LocastUser
This update will break previous OpenLocast-Web instances. South may not
be sufficient to migrate from the old user model to the new one... the
best way to do it might be to do a dumpdata (see: django-admin.py),
massage it to match the new user model, and then load it back in.